Update to official version of 3.60 api header. Doesn't change wrapper.

+++ 3.60/virtuoseAPI.h
@@ -217,7 +217,7 @@ extern "C"
 
   } virtDLLFunction;
 
-  extern int virtCallDLL(virtDLLFunction, ...);
+  extern size_t virtCallDLL(virtDLLFunction, ...);
   extern int virtLoadDLL(char *path);
